Move unit test back to Playwright
