--- .github/workflows/vitest.yml
@@ -1,22 +1,0 @@
-name: Vitest
-on: [push, pull_request]
-jobs:
-  test:
-    if: github.event_name != 'pull_request' || github.event.pull_request.head.repo.full_name != github.repository
-    timeout-minutes: 15
-    runs-on: ubuntu-24.04
-    steps:
-      - uses: actions/checkout@v4
-
-      - uses: actions/setup-node@v4
-        with:
-          node-version: 22.14
-
-      - name: Build Inertia
-        run: |
-          npm install
-          cd packages/core
-          npm run build
-
-      - name: Run Vitest Tests
-        run: cd packages/core && npm run test
