+++ packages/core/package.json
@@ -62,7 +62,6 @@
     "@types/qs": "^6.9.0",
     "esbuild": "^0.25.0",
     "esbuild-node-externals": "^1.6.0",
-    "typescript": "^4.9.4",
-    "vitest": "^3.2.4"
+    "typescript": "^4.9.4"
   }
 }
